- Caller of [APIHandler getResponse] receives an "id" as return value; - BDC sendInvoice API change. - Added GET API call for BDC ui API's: ui/ApproverList.json & ui/CreateBillApprover.json - Temporarily open all sections to all roles, need to fix later based on each role.

+++ BDC/SplashViewController.m
@@ -114,7 +114,7 @@
                 
                 [[Organization getSelectedOrg] getOrgFeatures];
             } else {
-                Error(@"Splash View Controller: Failed to login! %d, %@", status, err);
+                Error(@"Splash View Controller: Failed to login! %ld, %@", (long)status, err);
                 [self backToLogin];
             }
         }];
